+++ src/main/java/net/anotheria/anoprise/cache/RoundRobinSoftReferenceExperimentalCache.java
@@ -77,8 +77,8 @@ public class RoundRobinSoftReferenceExperimentalCache<K,V> extends AbstractCache
             id2index.remove(id);
             index2id.remove(index);
 
-            SoftReference<V> ref = cache[index.intValue()];
-            cache[index.intValue()] = null;
+            SoftReference<V> ref = cache[index];
+            cache[index] = null;
 
             if (ref.get() == null) {
                 cacheStatsCopy.addGarbageCollected();
@@ -100,7 +100,7 @@ public class RoundRobinSoftReferenceExperimentalCache<K,V> extends AbstractCache
             //System.out.println("Index for id: "+index);
             if (index == null)
                 return null;
-            V toRet = cache[index.intValue()].get();
+            V toRet = cache[index].get();
 
             if (toRet == null) {
                 cacheStatsCopy.addGarbageCollected();
@@ -130,14 +130,14 @@ public class RoundRobinSoftReferenceExperimentalCache<K,V> extends AbstractCache
             Integer oldPosition = getCachePosition(id);
             if (oldPosition != null) {
                 //System.out.println("replacing old entry.");
-                cache[oldPosition.intValue()] = toPut;
+                cache[oldPosition] = toPut;
                 return;
             }
 
             if (lastElement < currentSize - 1 && !firstCycleComplete) {
                 //es ist noch platz im array.
                 lastElement++;
-                Integer tmpIndex = Integer.valueOf(lastElement);
+                Integer tmpIndex = lastElement;
                 index2id.put(tmpIndex, id);
                 id2index.put(id, tmpIndex);
                 cache[lastElement] = toPut;
@@ -149,7 +149,7 @@ public class RoundRobinSoftReferenceExperimentalCache<K,V> extends AbstractCache
                 lastElement++;
                 if (lastElement == cache.length)
                     lastElement = 0;
-                Integer tmp = Integer.valueOf(lastElement);
+                Integer tmp = lastElement;
                 Object oldId = index2id.get(tmp);
                 if (oldId != null) {
                     id2index.remove(oldId);
